+++ src/grandorgue/config/GOConfig.h
@@ -189,6 +189,8 @@ public:
     GOMidiObject::ObjectType type, unsigned index);
   // search among user-added MIDI objects
   GOConfigMidiObject *FindMidiInitialObject(const wxString &path);
+  void AssignToInitial(const GOMidiObject &pObjFrom);
+  void DelMidiInitial(unsigned index);
 
   const std::vector<wxString> &GetAudioGroups();
   void SetAudioGroups(const std::vector<wxString> &audio_groups);